+++ src/movies.html
@@ -63,6 +63,6 @@
         </footer>
 
     <div id="root"></div>
-    <script src="main.js" type="module"></script>
+    <script src="movies.js" type="module"></script>
   </body>
 </html>
